backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
bcrypt
python-jose[cryptography]
python-dotenv
sqladmin